+++ CHANGELOG.md
@@ -15,7 +15,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - `LiftA2` to lift and apply a `Bifunction` to two `Applicative`s
 - `Flatten` to lazily flatten nested `Iterable<Iterable<A>>`s to `Iterable<A>`
 - `Replicate`, short-hand composition of `take` and `repeat`
-- `Distinct` to produce an `Iterable` of distinct values in another `Iterable` 
+- `Distinct` to produce an `Iterable` of distinct values in another `Iterable`
+- `Sort` and `SortBy` for eagerly, monolithically sorting `Iterable`s and producing `List`s  
 
 ## [1.6.2] - 2017-08-20
 ### Fixed